backend/requirements.txt
streamlit
requests
pandas
matplotlib
beautifulsoup4
fpdf
google-generativeai
langgraph
